requirements.txt
asgiref==3.2.7
Django==3.0
djangorestframework==3.10.0
djangorestframework-stubs==1.1.0
itsdangerous==1.1.0
mysqlclient==1.3.12
Pillow==7.1.1
pytz==2019.3
sqlparse==0.3.1
view==0.1
